+++ Modules/Intent.Modules.ModuleBuilder.CSharp/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 3.6.1
 
+- Improvement: Added support for NuGet Package dependencies.
 - Improvement: Fixed bug around using generation of NugetPackages template.
 
 ### Version 3.6.0